Move set_image and set_gradient into nsStyleImage impl
